refactor(renderer): share URL-scheme allowlist via schema; src delegates
